requirements.txt
bidict==0.21.2
click==8.0.1
Flask==2.0.1
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.3.2
PyMySQL==1.0.2
python-dotenv==0.19.0
python-engineio==4.2.1
python-socketio==5.4.0
SQLAlchemy==1.4.23
Werkzeug==2.0.1
WTForms==2.3.3
flask_wtf
